+++ scenes/timeclock_scene_menu.c
@@ -12,6 +12,7 @@ typedef enum {
     MenuIndexWeek,
     MenuIndexExport,
     MenuIndexSettings,
+    MenuIndexAbout,
 } MenuIndex;
 
 static void timeclock_scene_menu_submenu_callback(void* context, uint32_t index) {
@@ -41,6 +42,8 @@ void timeclock_scene_menu_on_enter(void* context) {
         submenu, "Export", MenuIndexExport, timeclock_scene_menu_submenu_callback, app);
     submenu_add_item(
         submenu, "Settings", MenuIndexSettings, timeclock_scene_menu_submenu_callback, app);
+    submenu_add_item(
+        submenu, "About", MenuIndexAbout, timeclock_scene_menu_submenu_callback, app);
 
     submenu_set_selected_item(
         submenu, scene_manager_get_scene_state(app->scene_manager, TimeClockSceneMenu));
@@ -57,6 +60,7 @@ bool timeclock_scene_menu_on_event(void* context, SceneManagerEvent event) {
         consumed = true;
         switch(event.event) {
         case MenuIndexPunch:
+            app->scan_purpose = TcScanPunch;
             scene_manager_next_scene(app->scene_manager, TimeClockSceneScan);
             break;
         case MenuIndexWork:
@@ -76,6 +80,9 @@ bool timeclock_scene_menu_on_event(void* context, SceneManagerEvent event) {
             break;
         case MenuIndexSettings:
             scene_manager_next_scene(app->scene_manager, TimeClockSceneSettings);
+            break;
+        case MenuIndexAbout:
+            scene_manager_next_scene(app->scene_manager, TimeClockSceneAbout);
             break;
         default:
             consumed = false;